gpixmap: use a single threshold for is_a_candidate_for_accel()
Preparing support to configure a custom threshold in the future.

The threshold is measured in bytes, occupied by the surface

NOTE: this is probably how it was intended, having the same
threshold in bytes for 8 and 32 bpp pixmaps.
The original thresholds seem a bit odd, pixelsize 12000 for 8 bpp,
vs 48000 for 32 bpp.
When doing cpu fill or blit, the total time spent is more likely to be
proportional to the number of bytes to be copied or cleared than to
the number of pixels.
Also, accelerated operations on 32bpp surfaces are not likely to be
4 times faster than the same operations on 8bpp surfaces, for the
same number of pixels.
And finally, because the memory available for accelerated surfaces is
limited, it makes sense to use the same size limit for 8 bpp as for
32 bpp pixmaps.

So use a single threshold in bytes, for both 8 bpp and 32 bpp pixmaps.