[Setup] Improve portability & speed of children check in setupDom()
In checkItems(), change the check on whether children are permitted
for an element to be more portable between Py2 & Py3 (and about 4
times faster) than the previous code.

The change in the test avoids an expected exception on all elements
that should have no children and in fact have no children.