[Timeshift.py] Refactor code to sub class Setup
This update should be functionally the same as the previous version except that rather than copying parts of the "Setup" code this version uses the "Setup" code directly.

There is one functional difference in that if the timeshift directory is invalid, using the Flash, read only media or doesn't support hard links, the GREEN button will be disabled and the user can not leave the setting entry field until a valid entry is selected.