[InfoBarGenerics] Fix a localization issue
Confusion has been caused because ConfigSelection has the value as item[0] and the displayed text as item[1], and ChoiceBox has the displayed text as item[0] and the value as item[1]... so passing the values between the ConfigSelection and the ChoiceBox has become confused.

And obviously, developers working in English would not notice the problem.