[enigma2.sh.in] reinstate code removed in 456cc4b931e160ff5fc9ecf348d0c6e5259df35f
Without this enigma will run slow when debug logs are not enabled (due to the large quantity of output flooding the serial port).

With this commit output is sent to the serial port on first install. After that serial port output is not needed as we can read from debug logs.