[Navigation.py] Tweak FCC handling
Do not set self.currentlyPlayingServiceReference and self.currentlyPlayingServiceOrGroup to None in the dispatchEvent if the service event is a zap.

Fixes ChannelNumber "---" bug, including removing unnecessary self.pnav.stopService() when zapping, which improves zapping performance.

This bug has taken a lot of time and hard work to track down.

Co-authored with @dpeddi from BlackHole