[RecordTimer] Fix logic error in cleanup()
When record timer remove callbacks were added to RecordTimer.cleanup(),
the condition in constructing the removed_timers list should have
had "not" rather than "not not".

The error happened when the changes were rewritten from the original
to allow merging with OpenViX, where the cleanup() method was already
defined, while it was not in the original Beyonwiz code.

The commit removes the unwanted extra "not".