Add possibility to indent configs in setup.xml files
Add a possiblity to indent a config in the setup.xml file by adding an
indent attribute where you can put a string (most likelu spaces) to
indent a config. E.g. when you have a switch or a config that allows
more additional settings then it can indent so you can build a better
overview. The new attribute in setup.xml is just a simple string that
will be added in front of the config text.