Revert "Enable parent ".." in default path (will still be disabled if it is also the root). (#1119)"
This reverts commit f7a1420c92ba9fde6063a66b0c84b1d8476a988f.

@BrianG61UK commented on Apr 14, 2025
Is this why it's now possible to go into .. in the movie folder?
If so, why on Earth?
We don't want users to be able to play anything other than movies.
We don't want users to be able to delete anything other than movies.
This seems like a disaster to me.

@RobvanderDoes commented on Apr 14, 2025
During tests under 6.7.008.004 it worked fine (i.e. not going to 'deep'). No idea why it does now on the current dev/release image.

@Huevos commented on Apr 14, 2025
Reverted because as I already said it should not allow users outside the base movie folder. The current implementation allows things such as sending your picon folder to the movie trashcan.